Velocity analisys with velocity continuation
sfcosft is the Multi-dimensional cosine transform. Pad the data
if you need to suppress wrap-around effects. The signal 1 is the foward
trasform, the -1 is inverse.

sfstolt is the post-stack Stolt migration. And it requires the input to
be cosine-transformed over the lateral axes.

sfvczo is the zero offset velocity continuation. The data is migrated
multiple times with a different constant velocity

In this step, two velocity datas are generated, one of them with the
amplitude samples squared and they will be used to get velocity panel
x focusing measure