A study of the recipe in 'stack.py' script #3
The original script, 'stack.py', is a Madagascar recipe for stacking and
diffraction imaging. So, 'stack.study.py' is a study of this script with
coments and explanation of what each function does.

Firts it receives a seimic data cube in the function stack, do the
conventional NMO stacking and DMO stacking and pass the data for the
diffimg function to migrate the diffractions.

It separates reflections and diffractions using plane wave destruction
filters, do the velocity analisys by focusing measure and migrate them.