Scratch of Mdiffsim.c to simulate diffraction #5
The madagascar program 'Mdiffsim.c' is the scratch of the structure of
the program to simulate diffraction hyperbolas in the stacked section.
This first version only receives the stacked section and read the
parameters in stdin.

The usage example in the usage_example/sfdiffsim is just a test of this
program. It generates a copy of the input (the stacked section)