Ricker convolution in hyperbola coordinates Resolve #7
Use Madagascar API's to generate a ricker pulse in a trace vector
called 'ricker'. It is generated with a convolution between a spike
with a ricker filter using ricker.h library.

To generate the hyperbola, just shift the ricker wavelet in time, puting
it's center (max amplitude) in the time coordinate determined by the calculated
hyperbola time curve t(m), and sum up with the amplitude in the stackedSection.