Modify Mdiffsim.c to output diffraction section #9
The purpose of diffraction simulation process is to colapse the diffraction
hyperbolas and discover the best velocity to do it. So, there is no need
for reflections. Stacked section is only use for picking of hyperbolas
center coordinates.

So, the program is modified to output a section with only diffraction
hyperbolas, and migrate it using focusing measure.