Modify velocity input to several reflectors #17
Program sfdiffsim now reads a velocity vector with a sample for each
reflector. Each sample of the velocity vector is assigned to a set of
(t0,m0) pairs in t0s and m0s files. Because of that, t0s and m0s files
now have n2 axis assigned to a given velocity sample in velocity vector.

So, v[0] corresponds to a velocity sample for
(t0[0][0],m0[0][0]),(t0[1][0],m0[1][0]),(t0[2][0],m0[2][0])... pairs

and  v[1] corresponds to a velocity sample for
(t0[0][1],m0[0][1]),(t0[1][1],m0[1][1]),(t0[2][1],m0[2][1])... pairs

and so on...